+++ src/sunflare/engine/_wrapper.py
@@ -2,11 +2,9 @@
 
 from __future__ import annotations
 
-import asyncio
-import sys
 from concurrent.futures import Future, ThreadPoolExecutor
 from itertools import count
-from typing import Any, Callable, Optional, Union
+from typing import Any, Callable
 
 import zmq
 from bluesky.run_engine import (
@@ -32,21 +30,21 @@ from sunflare.virtual import encode
 
 __all__ = ["RunEngine", "RunEngineResult", "RunEngineInterrupted"]
 
-DocumentType = Union[
-    RunStart,
-    RunStop,
-    EventDescriptor,
-    Event,
-    EventPage,
-    Datum,
-    DatumPage,
-    Resource,
-    StreamResource,
-    StreamDatum,
-]
+DocumentType = (
+    RunStart
+    | RunStop
+    | EventDescriptor
+    | Event
+    | EventPage
+    | Datum
+    | DatumPage
+    | Resource
+    | StreamResource
+    | StreamDatum
+)
 
-REResultType = Union[RunEngineResult, tuple[str, ...], Exception]
-FuncSocket = Union[Callable[[str, dict[str, Any]], None], zmq.Socket[bytes]]
+REResultType = RunEngineResult | tuple[str, ...] | Exception
+FuncSocket = Callable[[str, dict[str, Any]], None] | zmq.Socket[bytes]
 
 _prefix_counter = count()
 
@@ -98,20 +96,13 @@ class RunEngine(BlueskyRunEngine):
         self.socket_prefix: str = kwargs.pop(
             "socket_prefix", f"RE{next(_prefix_counter)}"
         )
-        self.socket: Optional[zmq.Socket[bytes]] = kwargs.pop("socket", None)
+        self.socket: zmq.Socket[bytes] | None = kwargs.pop("socket", None)
         self._executor = ThreadPoolExecutor(max_workers=1)
         self._result: REResultType
         super().__init__(*args, **kwargs)
 
         # override pause message to be an empty string
         self.pause_msg = ""
-
-        # Python 3.9 explicitly requires the event loop to be set
-        # when running in a separate thread
-        if sys.version_info < (3, 10):
-            self._run_in_executor = self.__run_in_executor_explicit
-        else:
-            self._run_in_executor = self.__run_in_executor
 
     def emit_sync(self, name: DocumentNames, doc: dict[str, Any]) -> None:
         """Emit a document synchronously.
@@ -131,16 +122,14 @@ class RunEngine(BlueskyRunEngine):
             self.socket.send_multipart([topic.encode(), encode(doc)])
         super().emit_sync(name, doc)
 
-    def __run_in_executor_explicit(self, *args: Any, **kwargs: Any) -> REResultType:
-        asyncio.set_event_loop(self.loop)
-        return super().__call__(*args, **kwargs)  # type: ignore[no-any-return]
-
     def __run_in_executor(self, *args: Any, **kwargs: Any) -> REResultType:
         return super().__call__(*args, **kwargs)  # type: ignore[no-any-return]
 
     def __call__(self, *args: Any, **metadata_kw: Any) -> Future[REResultType]:
         self._fut = self._executor.submit(
-            lambda: self._run_in_executor(*args, **metadata_kw)
+            super().__call__,
+            *args,
+            **metadata_kw,
         )
         self._fut.add_done_callback(self._set_result)
         return self._fut